requirements.txt
fastapi
uvicorn[standard]==0.23.2
pydantic
pydantic_settings
discord.py
prisma
discord.py
python-dotenv==1.0.1
psycopg2-binary==2.9.9
types-requests
# GraphQL (Added)
strawberry-graphql[fastapi,pydantic]

omise

# Linters / type checker / formatter
ruff>=0.3.0
mypy>=1.8.0
typeguard
